package/server/requirements.txt
# FastAPI
fastapi==0.122.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
SQLAlchemy==2.0.44
alembic==1.17.2
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Pydantic & Settings
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.2.0

# Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4

# HTTP Client
aiohttp==3.11.14
httpx==0.28.1

# Scheduling
APScheduler==3.11.0

# Image Processing
Pillow==11.1.0

# Data Processing
pandas==2.2.3
numpy==2.2.3

# Utilities
python-dotenv==1.0.1
pytz==2025.1
typing-extensions==4.12.2

# Development
pytest==8.3.5
pytest-asyncio==0.25.3
black==25.1.0
isort==6.0.1
mypy==1.15.0
